Split most map object selection code to separate file
