+++ appOPHD/States/MapViewStateMapObjectPicker.cpp
@@ -1,0 +1,105 @@
+#include "MapViewState.h"
+
+#include "MapViewStateHelper.h"
+#include "../PointerType.h"
+#include "../MapObjects/RobotType.h"
+
+
+bool MapViewState::isInserting() const
+{
+	return mInsertMode != InsertMode::None;
+}
+
+
+bool MapViewState::isInsertingRobot() const
+{
+	return mInsertMode == InsertMode::Robot;
+}
+
+
+/**
+ * Clears the build mode.
+ */
+void MapViewState::clearMode()
+{
+	mInsertMode = InsertMode::None;
+	setCursor(PointerType::Normal);
+
+	mCurrentStructure = StructureID::SID_NONE;
+	mCurrentRobot = RobotType::None;
+}
+
+
+void MapViewState::onMouseWheel(NAS2D::Vector<int> changeAmount)
+{
+	if (mInsertMode != InsertMode::Tube) { return; }
+
+	changeAmount.y > 0 ? mConnections.decrementSelection() : mConnections.incrementSelection();
+}
+
+
+void MapViewState::onStructuresSelectionChange(const IconGridItem* item)
+{
+	if (!item)
+	{
+		clearMode();
+		return;
+	}
+
+	mConnections.clearSelection();
+	mRobots.clearSelection();
+
+	const auto structureId = static_cast<StructureID>(item->meta);
+
+	// Check availability
+	if (!item->available)
+	{
+		resourceShortageMessage(mResourcesCount, structureId);
+		mStructures.clearSelection();
+		return;
+	}
+
+	mCurrentStructure = structureId;
+	mInsertMode = InsertMode::Structure;
+	setCursor(PointerType::PlaceTile);
+}
+
+
+/**
+ * Handler for the Tubes Pallette dialog.
+ */
+void MapViewState::onConnectionsSelectionChange(const IconGridItem* item)
+{
+	if (!item)
+	{
+		clearMode();
+		return;
+	}
+
+	mRobots.clearSelection();
+	mStructures.clearSelection();
+
+	mCurrentStructure = StructureID::SID_TUBE;
+	mInsertMode = InsertMode::Tube;
+	setCursor(PointerType::PlaceTile);
+}
+
+
+/**
+ * Handles clicks of the Robot Selection Menu.
+ */
+void MapViewState::onRobotsSelectionChange(const IconGridItem* item)
+{
+	if (!item)
+	{
+		clearMode();
+		return;
+	}
+
+	mConnections.clearSelection();
+	mStructures.clearSelection();
+
+	mCurrentRobot = static_cast<RobotType>(item->meta);
+	mInsertMode = InsertMode::Robot;
+	setCursor(PointerType::PlaceTile);
+}
